refactor(typechecker): remove content-mapper croquis naming dependency
